requirements.txt
openai>=1.0.0
langgraph>=0.0.50
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
langfuse>=2.30.0
